server/requirements.txt
Flask
pyjwt
bcrypt
